move ComponentRegistry to pro directory
